refactor out trf
