+++ .vscode/settings.json
@@ -170,6 +170,7 @@
     "genindex",
     "geos",
     "gmail",
+    "gridspec",
     "holger",
     "howto",
     "hstack",
